CodeCraftAgent/backend/requirements.txt
# backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
aiohttp==3.10.0
python-dotenv==1.0.1
black==24.8.0
flake8==6.0.0
pytest==7.4.0